code simplified and working

+++ src/interfaces/IShortestPath.java
@@ -7,5 +7,5 @@ import dataStructures.GPSSignal;
 public interface IShortestPath extends ITrip {
 	public void addInstance(GPSSignal s, Integer speedLimit);
 	public Vector<Integer> getSpeedLimits();
-	public Integer getSpeedLimitAt(Integer i);
+	public double getSpeedLimitAt(Integer i);
 }
